API/requirements.txt
uv
hvac
PyJWT
pyodbc
sentence-transformers
python-multipart
pinecone-client
ptvsd
google-search-results
numexpr
huggingface_hub
protobuf
sqlalchemy
sqlalchemy.orm
pandas
langgraph
langchain_openai
python-dotenv==1.0.1
openai==1.16.0
langchain==0.1.11
langchain-community==0.0.25
langsmith==0.1.19
psycopg2-binary==2.9.9
tiktoken==0.5.1
pgvector==0.2.5
urllib3==2.0.3
pip-system-certs==4.0
fastapi==0.110.0
pydantic==2.6.3
uvicorn==0.29.0
IPython
